[VEUE-192] abstracts copyToClipboard into helper method; changes css to use class-based styling for button (.copy-to-clipboard)
