+++ app/javascript/controllers/broadcast/browser_controller.ts
@@ -32,7 +32,6 @@ export default class extends Controller {
       this.updateState(navigationUpdate);
     };
     ipcRenderer.on("browserView", this.browserViewListener);
-
     this.navigateToAddress();
   }
